Add a dropdown menu to the gvdw interactive notebook to select different potentials:
* Lennard-Jones
* Debye Huckel
* Or both

I was not particularly happy about this line:

```python
parameters = {k:v for k,v in parameters.items() if k in inspect.signature(potential).parameters}\n",
```

It inspects the function object for which parameters it needs and then provides them from the one that are available from the sliders. It works, but it is not elegant.